requirements.txt
a2wsgi==1.10.8
aiologic==0.14.0
aiosqlite==0.21.0
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
apache-airflow==3.0.0
apache-airflow-core==3.0.0
apache-airflow-providers-celery==3.10.6
apache-airflow-providers-common-compat==1.6.0
apache-airflow-providers-common-io==1.5.4
apache-airflow-providers-common-sql==1.26.0
apache-airflow-providers-smtp==2.0.3
apache-airflow-providers-standard==1.0.0
apache-airflow-task-sdk==1.0.0
argcomplete==3.6.2
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
cadwyn==5.3.3
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorlog==6.9.0
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==42.0.8
Deprecated==1.2.18
dill==0.3.1.1
dnspython==2.7.0
docopt==0.6.2
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
Flask==2.2.5
flower==2.0.1
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
gunicorn==23.0.0
h11==0.14.0
hdfs==2.7.3
httpcore==1.0.8
httptools==0.6.4
httpx==0.27.0
humanize==4.12.2
idna==3.10
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kafka-python==2.1.5
kombu==5.5.3
lazy-object-proxy==1.11.0
libcst==1.7.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.7.0
msgspec==0.19.0
numpy==2.2.4
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pendulum==3.1.0
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
protobuf==4.25.6
psutil==7.0.0
py4j==0.10.9.7
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
pyspark==3.5.5
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
retryhttp==1.3.2
rich==13.9.4
rich-argparse==1.7.0
rich-toolkit==0.14.1
rpds-py==0.24.0
setproctitle==1.3.5
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
starlette==0.46.2
structlog==25.2.0
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.0.1
text-unidecode==1.3
tornado==6.4.2
typer==0.15.2
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
universal_pathlib==0.2.6
urllib3==2.4.0
uuid6==2024.7.10
uvicorn==0.34.2
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==2.2.3
wirerope==1.0.0
wrapt==1.17.2
zipp==3.21.0
